Move components for auto labeling config
