+++ frontend/components/configAutoLabeling/ConfigCreationForm.vue
