+++ frontend/components/configAutoLabeling/form/ConfigHeader.vue
