+++ frontend/components/configAutoLabeling/form/ConfigLabelMapping.vue
