+++ frontend/components/configAutoLabeling/form/ConfigParameters.vue
