+++ frontend/components/configAutoLabeling/form/ConfigTemplate.vue
