+++ frontend/components/configAutoLabeling/form/ConfigTemplateName.vue
