+++ frontend/components/configAutoLabeling/form/LabelMapping.vue
