+++ frontend/components/configAutoLabeling/form/ObjectField.vue
